task 6

+++ 0x02-functions_nested_loops/6-abs.c
@@ -11,7 +11,7 @@ int _abs(int n)
 {
 if (n < 0)
 {
-n *= -1/10;
+n *= double(-1/10);
 printf("%d", n);
 }
 else if (n >= 0)
